+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feature: install déjà dup backups
 - feature: install draw.io
 - feature: install dropbox
+- feature: install obsidian
 
 ## [1.4.0] - 2026-07-24